Upload June 19
This is the dump of the data and procedures that I have so far.  I have not finished adding all of the ranged weapon data as well as some other data, but this is a good starting place.

The script file should be unnecessary to run, but I wanted to include it so you can see all of the procedures I've built so far and critique what you think needs changing.